Creating Expense resource

+++ src/main/java/br/com/mybudget/usermanager/model/dto/ExpenseEnvelopeDTO.java
@@ -17,6 +17,6 @@ import lombok.ToString;
 @ToString
 public class ExpenseEnvelopeDTO {
 	
-	private int idUser;
+	private Long idUser;
 	private List<ExpenseRequestDTO> expenses;
 }
